requirements.txt
Django>=2.2.4
mysqlclient==1.4.6
google-cloud-storage==1.23.0
